Redone all Shipment related DAL classes and DTOs

+++ dev/Assignment1/src/DTO/ItemDTO.java
@@ -3,9 +3,9 @@ package DTO;
 public class ItemDTO {
     private int id;
     private String name;
-    private double weight;
+    private float weight;
 
-    public ItemDTO(int id, String name, double weight) {
+    public ItemDTO(int id, String name, float weight) {
         this.id = id;
         this.name = name;
         this.weight = weight;
@@ -19,7 +19,7 @@ public class ItemDTO {
         return name;
     }
 
-    public double getWeight() {
+    public float getWeight() {
         return weight;
     }
